requirements.txt
asgiref==3.7.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.2
django-jazzmin==2.6.0
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper==20230314
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
